feat(season): register mongoose change tracking plugin to schema

+++ src/routes/season.routes.ts
@@ -1,4 +1,5 @@
 import express from "express";
+import { DecodedIdToken } from "firebase-admin/auth";
 import Season from "../models/season.model";
 import { isAdmin } from "../utils/middlewares/middlewares";
 
@@ -25,10 +26,12 @@ seasonRoutes.get("/", async (req, res, next) => {
 
 seasonRoutes.post("/", isAdmin(), async (req, res, next) => {
   try {
+    const token: DecodedIdToken = res.locals.token;
+
     const id = await Season.fromJsonApi(req.body, {
       assignAttribute: Season.fromLanguage(req.query.language),
     })
-      .save()
+      .save({ user: token.uid })
       .then((doc) => doc._id);
 
     const response = await Season.findById(id)
@@ -61,6 +64,8 @@ seasonRoutes.get("/:id", async (req, res, next) => {
 
 seasonRoutes.patch("/:id", isAdmin(), async (req, res, next) => {
   try {
+    const token: DecodedIdToken = res.locals.token;
+
     await Season.findById(req.params.id)
       .orFail()
       .then((doc) => {
@@ -68,7 +73,7 @@ seasonRoutes.patch("/:id", isAdmin(), async (req, res, next) => {
           .merge(Season.fromJsonApi(req.body, {
             assignAttribute: Season.fromLanguage(req.query.language),
           }))
-          .save();
+          .save({ user: token.uid });
       });
 
     const response = await Season.findById(req.params.id)
@@ -86,11 +91,13 @@ seasonRoutes.patch("/:id", isAdmin(), async (req, res, next) => {
 
 seasonRoutes.delete("/:id", isAdmin(), async (req, res, next) => {
   try {
+    const token: DecodedIdToken = res.locals.token;
+
     await Season.findById(req.params.id)
       .orFail()
       .then((doc) => {
         return doc
-          .deleteOne();
+          .deleteOne({ user: token.uid });
       });
 
     res.status(204).send();
@@ -120,6 +127,26 @@ seasonRoutes.get("/:id/episodes", async (req, res, next) => {
   try {
     const response = await Season.findById(req.params.id)
       .getRelationship("episodes")
+      .withJsonApi(req.query)
+      .withLanguage(req.query.language)
+      .toJsonApi({
+        baseUrl: `${process.env.API_URL}`,
+      })
+      .paginate({
+        url: `${process.env.API_URL}${req.originalUrl}`,
+        query: req.query,
+      });
+
+    res.json(response);
+  } catch (err) {
+    next(err);
+  }
+});
+
+seasonRoutes.get("/:id/changes", async (req, res, next) => {
+  try {
+    const response = await Season.findById(req.params.id)
+      .getRelationship("changes")
       .withJsonApi(req.query)
       .withLanguage(req.query.language)
       .toJsonApi({
